requirements.txt
fastapi
uvicorn
Pillow
pillow
torch
torchvision
requests
pytest
flake8 
pylint 
black
bandit
pip-audit